requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
djangorestframework==3.7.1
gunicorn==19.9.0
html5lib==0.999999999
idna==2.6
numpy==1.14.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
